+++ hiddifypanel/hutils/proxy/xrayjson.py
@@ -8,7 +8,7 @@ from hiddifypanel.models import hconfig, ConfigEnum
 from .xray import is_muxable_agent, OUTBOUND_LEVEL
 
 
-def configs_as_json(domains: list[Domain], user: User, expire_days: int,  remarks: str) -> str:
+def configs_as_json(domains: list[Domain], user: User, expire_days: int, remarks: str) -> str:
     '''Returns xray configs as json'''
     all_configs = []
 
@@ -35,50 +35,48 @@ def configs_as_json(domains: list[Domain], user: User, expire_days: int,  remark
         )
     # endregion
 
-    # region show status (active/disable)
-    active = True
     if not user.is_active:
+        # region show status (active/disable)
         tag = '✖ ' + (hutils.encode.url_encode('بسته شما به پایان رسید') if hconfig(ConfigEnum.lang) == 'fa' else 'Package Ended')
         # add user status
         all_configs.append(
             null_config(tag)
         )
-        active = False
-    # endregion
-
-    if active:
+        # endregion
+    else:
         # TODO: seperate codes to small functions
         # TODO: check what are unsupported protocols in other apps
-        unsupported_protos = []
-        unsupported_transport = []
+        unsupported_protos = {}
+        unsupported_transport = {}
         if g.user_agent.get('is_v2rayng'):
             # TODO: ensure which protocols are not supported in v2rayng
-            unsupported_protos = [ProxyProto.wireguard, ProxyProto.hysteria, ProxyProto.hysteria2, ProxyProto.tuic, ProxyProto.ss, ProxyProto.ssr, ProxyProto.ssh]
+            unsupported_protos = {ProxyProto.wireguard, ProxyProto.hysteria, ProxyProto.hysteria2,
+                                  ProxyProto.tuic, ProxyProto.ss, ProxyProto.ssr, ProxyProto.ssh}
             if not hutils.flask.is_client_version(hutils.flask.ClientVersion.v2ryang, 1, 8, 18):
-                unsupported_transport = [ProxyTransport.httpupgrade]
+                unsupported_transport = {ProxyTransport.httpupgrade}
 
         # multiple outbounds needs multiple whole base config not just one with multiple outbounds (at least for v2rayng)
         # https://github.com/2dust/v2rayNG/pull/2827#issue-2127534078
         outbounds = []
         for proxy in hutils.proxy.get_valid_proxies(domains):
-            if unsupported_protos and proxy['proto'] in unsupported_protos:
+            if proxy['proto'] in unsupported_protos:
                 continue
-            if unsupported_transport and proxy['transport'] in unsupported_transport:
+            if proxy['transport'] in unsupported_transport:
                 continue
             outbound = to_xray(proxy)
-            if 'msg' not in outbound:
-                outbounds.append(outbound)
+            outbounds.append(outbound)
 
         base_config = json.loads(render_template('base_xray_config.json.j2', remarks=remarks))
         if len(outbounds) > 1:
             for out in outbounds:
-                base_config['remarks'] = out['tag']
-                base_config['outbounds'].insert(0, out)
-                if all_configs:
-                    all_configs.insert(0, copy.deepcopy(base_config))
-                else:
-                    all_configs.append(copy.deepcopy(base_config))
-                del base_config['outbounds'][0]
+                base = copy.deepcopy(base_config)
+                base['remarks'] = out['tag']
+                base['outbounds'].insert(0, out)
+                # if all_configs:
+                #     all_configs.insert(0, copy.deepcopy(base_config))
+                # else:
+                all_configs.append(base)
+
         else:  # single outbound
             base_config['outbounds'].insert(0, outbounds[0])
             all_configs = base_config