+++ hiddifypanel/panel/commercial/restapi/v2/admin/server_status_api.py
@@ -20,6 +20,7 @@ class AdminServerStatusApi(MethodView):
 
     @app.output(ServerStatusOutputSchema)  # type: ignore
     def get(self):
+        """System: ServerStatus"""
         dto = ServerStatusOutputSchema()
         dto.stats = {  # type: ignore
             'system': hutils.system.system_stats(),