+++ hiddifypanel/panel/commercial/restapi/v2/admin/user_api.py
@@ -9,7 +9,7 @@ from hiddifypanel.drivers import user_driver
 from hiddifypanel.panel import hiddify
 
 from . import has_permission
-from .schema import UserSchema, PutUserSchema, PatchUserSchema, SuccessfulSchema
+from .schema import UserSchema, PostUserSchema, PatchUserSchema, SuccessfulSchema
 
 
 class UserApi(MethodView):
@@ -17,30 +17,17 @@ class UserApi(MethodView):
 
     @app.output(UserSchema)  # type: ignore
     def get(self, uuid):
+        """User: Get details of a user"""
         user = User.by_uuid(uuid) or abort(404, "User not found")
         if not has_permission(user):
             abort(403, "You don't have permission to access this user")
 
         return user.to_schema()  # type: ignore
 
-    @app.input(PutUserSchema, arg_name="data")  # type: ignore
-    @app.output(SuccessfulSchema)  # type: ignore
-    def put(self, uuid, data):
-        if User.by_uuid(uuid):
-            abort(400, 'The user exists')
-        data['uuid'] = uuid
-
-        if not data.get('added_by_uuid'):
-            data['added_by_uuid'] = g.account.uuid
-
-        dbuser = User.add_or_update(**data) or abort(502, "Unknown issue: User is not added")
-        user_driver.add_client(dbuser)
-        hiddify.quick_apply_users()
-        return {'status': 200, 'msg': 'ok'}
-
     @app.input(PatchUserSchema, arg_name="data")  # type: ignore
-    @app.output(SuccessfulSchema)  # type: ignore
+    @app.output(UserSchema)  # type: ignore
     def patch(self, uuid, data):
+        """User: Update a user"""
         user = User.by_uuid(uuid) or abort(404, "user not found")
         if not has_permission(user):
             abort(403, "You don't have permission to access this user")
@@ -50,17 +37,20 @@ class UserApi(MethodView):
                 continue
             if field not in data:
                 data[field] = getattr(user, field)
-
+        data['old_uuid'] = uuid
+        user_driver.remove_client(user)
         dbuser = User.add_or_update(**data) or abort(502, "Unknown issue! User is not patched")
-        user_driver.add_client(dbuser)
+        if dbuser.is_active:
+            user_driver.add_client(dbuser)
         # the add_or_update doesn't update the uuid of User, so for now just delete old user after adding new
-        if user.uuid != data['uuid']:
-            user.remove()
+        # if user.uuid != data['uuid']:
+        #     user.remove()
         hiddify.quick_apply_users()
-        return {'status': 200, 'msg': 'ok'}
+        return dbuser.to_schema()
 
     @app.output(SuccessfulSchema)  # type: ignore
     def delete(self, uuid):
+        """User: Delete a User"""
         user = User.by_uuid(uuid) or abort(404, "user not found")
         if not has_permission(user):
             abort(403, "You don't have permission to access this user")