+++ hiddifypanel/panel/commercial/restapi/v2/panel/schema.py
@@ -5,3 +5,8 @@ from apiflask import Schema, fields
 class PanelInfoOutputSchema(Schema):
     version = fields.String(description="The panel version")
 # endregion
+
+
+class PongOutputSchema(Schema):
+    msg = fields.String(description="Pong Response")
+# endregion